git_create -> git_create_and_set_new_origin: Clarify
It was unclear how `hub` was handling remotes,
renaming, etc.

Try to document and explain what's happening. With
the rename of the former origin -> origin-old,
this is no longer clobbering the origin shortname.

This means that git_set_remote is _actually_
aligning the remote type based on the user's SSH
setting.